--- a/awx/ui/Gruntfile.js
+++ b/awx/ui/Gruntfile.js
@@ -17,38 +17,4 @@ module.exports = function(grunt) {
     grunt.initConfig(configs);
     grunt.loadNpmTasks('grunt-newer');
     grunt.loadNpmTasks('grunt-angular-gettext');
-
-    // writes environment variables for development. current manages:
-    // browser-sync + websocket proxy
-
-    grunt.registerTask('sync', [
-        'browserSync:http',
-        'concurrent:watch'
-    ]);
-
-    grunt.registerTask('dev', [
-        'clean:tmp',
-        'clean:static',
-        'concurrent:dev',
-        'concat:css',
-        'webpack:dev',
-        'sync'
-    ]);
-
-    grunt.registerTask('devNoSync', [
-        'clean:tmp',
-        'clean:static',
-        'concurrent:devNoSync',
-        'concat:css'
-    ]);
-
-    grunt.registerTask('release', [
-        'clean:tmp',
-        'clean:static',
-        'concurrent:prod',
-        'webpack:prod',
-        'concat:css',
-        'cssmin:vendor',
-        'cssmin:source'
-    ]);
 };