--- a/awx/ui/build/webpack.development.js
+++ b/awx/ui/build/webpack.development.js
@@ -1,0 +1,11 @@
+const path = require('path');
+
+const _ = require('lodash');
+
+let base = require('./webpack.base');
+
+let development = {
+    devtool: 'cheap-source-map'
+};
+
+module.exports = _.merge(base, development);