--- a/awx/ui/build/webpack.watch.js
+++ b/awx/ui/build/webpack.watch.js
@@ -7,7 +7,6 @@ const HtmlWebpackHarddiskPlugin = require('html-webpack-harddisk-plugin');
 const TARGET_PORT = _.get(process.env, 'npm_package_config_django_port', 8043);
 const TARGET_HOST = _.get(process.env, 'npm_package_config_django_host', 'https://localhost');
 const TARGET = `https://${TARGET_HOST}:${TARGET_PORT}`;
-const STATIC_URL = '/static/';
 
 let development = require('./webpack.development');
 
@@ -18,8 +17,8 @@ let watch = {
     ],
     devServer: {
         contentBase: path.resolve(__dirname, '..', 'static'),
-        publicPath: STATIC_URL,
         clientLogLevel: 'info',
+        publicPath: '/static/',
         host: '127.0.0.1',
         port: 3000,
         proxy: {