--- a/awx/ui/client/features/credentials/index.js
+++ b/awx/ui/client/features/credentials/index.js
@@ -3,6 +3,8 @@ import AddController from './add-credentials.controller';
 import EditController from './edit-credentials.controller';
 import CredentialsStrings from './credentials.strings'
 
+const addEditTemplate = require('@features/credentials/add-edit-credentials.view.html');
+
 function CredentialsResolve ($q, $stateParams, Me, Credential, CredentialType, Organization) {
     let id = $stateParams.credential_id;
 
@@ -69,7 +71,7 @@ function CredentialsConfig ($stateExtenderProvider, legacyProvider, pathProvider
         },
         views: {
             'add@credentials': {
-                templateUrl: path.getViewPath('credentials/add-edit-credentials'),
+                templateUrl: addEditTemplate,
                 controller: AddController,
                 controllerAs: 'vm'
             }
@@ -92,7 +94,7 @@ function CredentialsConfig ($stateExtenderProvider, legacyProvider, pathProvider
         },
         views: {
             'edit@credentials': {
-                templateUrl: path.getViewPath('credentials/add-edit-credentials'),
+                templateUrl: addEditTemplate,
                 controller: EditController,
                 controllerAs: 'vm'
             }