--- a/awx/ui/client/lib/components/index.js
+++ b/awx/ui/client/lib/components/index.js
@@ -1,39 +1,35 @@
-import layout from './layout/layout.directive';
-import topNavItem from './layout/top-nav-item.directive';
-import sideNav from './layout/side-nav.directive';
-import sideNavItem from './layout/side-nav-item.directive';
-import actionGroup from './action/action-group.directive';
-import divider from './utility/divider.directive';
-import form from './form/form.directive';
-import formAction from './form/action.directive';
-import inputCheckbox from './input/checkbox.directive';
-import inputGroup from './input/group.directive';
-import inputLabel from './input/label.directive';
-import inputLookup from './input/lookup.directive';
-import inputMessage from './input/message.directive';
-import inputSecret from './input/secret.directive';
-import inputSelect from './input/select.directive';
-import inputText from './input/text.directive';
-import inputTextarea from './input/textarea.directive';
-import inputTextareaSecret from './input/textarea-secret.directive';
-import modal from './modal/modal.directive';
-import panel from './panel/panel.directive';
-import panelHeading from './panel/heading.directive';
-import panelBody from './panel/body.directive';
-import popover from './popover/popover.directive';
-import tab from './tabs/tab.directive';
-import tabGroup from './tabs/group.directive';
-import truncate from './truncate/truncate.directive';
+import actionGroup from '@components/action/action-group.directive';
+import divider from '@components/utility/divider.directive';
+import form from '@components/form/form.directive';
+import formAction from '@components/form/action.directive';
+import inputCheckbox from '@components/input/checkbox.directive';
+import inputGroup from '@components/input/group.directive';
+import inputLabel from '@components/input/label.directive';
+import inputLookup from '@components/input/lookup.directive';
+import inputMessage from '@components/input/message.directive';
+import inputSecret from '@components/input/secret.directive';
+import inputSelect from '@components/input/select.directive';
+import inputText from '@components/input/text.directive';
+import inputTextarea from '@components/input/textarea.directive';
+import inputTextareaSecret from '@components/input/textarea-secret.directive';
+import layout from '@components/layout/layout.directive';
+import modal from '@components/modal/modal.directive';
+import panel from '@components/panel/panel.directive';
+import panelBody from '@components/panel/body.directive';
+import panelHeading from '@components/panel/heading.directive';
+import popover from '@components/popover/popover.directive';
+import sideNav from '@components/layout/side-nav.directive';
+import sideNavItem from '@components/layout/side-nav-item.directive';
+import tab from '@components/tabs/tab.directive';
+import tabGroup from '@components/tabs/group.directive';
+import topNavItem from '@components/layout/top-nav-item.directive';
+import truncate from '@components/truncate/truncate.directive';
 
-import BaseInputController from './input/base.controller';
-import ComponentsStrings from './components.strings';
+import BaseInputController from '@components/input/base.controller';
+import ComponentsStrings from '@components/components.strings';
 
 angular
     .module('at.lib.components', [])
-    .directive('atLayout', layout)
-    .directive('atTopNavItem', topNavItem)
-    .directive('atSideNav', sideNav)
-    .directive('atSideNavItem', sideNavItem)
     .directive('atActionGroup', actionGroup)
     .directive('atDivider', divider)
     .directive('atForm', form)
@@ -48,13 +44,17 @@ angular
     .directive('atInputText', inputText)
     .directive('atInputTextarea', inputTextarea)
     .directive('atInputTextareaSecret', inputTextareaSecret)
+    .directive('atLayout', layout)
     .directive('atModal', modal)
     .directive('atPanel', panel)
-    .directive('atPanelHeading', panelHeading)
     .directive('atPanelBody', panelBody)
+    .directive('atPanelHeading', panelHeading)
     .directive('atPopover', popover)
+    .directive('atSideNav', sideNav)
+    .directive('atSideNavItem', sideNavItem)
     .directive('atTab', tab)
     .directive('atTabGroup', tabGroup)
+    .directive('atTopNavItem', topNavItem)
     .directive('atTruncate', truncate)
-    .service('ComponentsStrings', ComponentsStrings)
-    .service('BaseInputController', BaseInputController);
+    .service('BaseInputController', BaseInputController)
+    .service('ComponentsStrings', ComponentsStrings);