… awx/ui/build/webpack.test.js
… awx/ui/client/index.template.html
… awx/ui/client/legacy/styles/angular-scheduler.less
… awx/ui/client/legacy/styles/animations.less
… awx/ui/client/legacy/styles/ansible-ui.less
… awx/ui/client/legacy/styles/bootstrap-datepicker.less
… awx/ui/client/legacy/styles/codemirror.less
… awx/ui/client/legacy/styles/dashboard.less
… awx/ui/client/legacy/styles/event-viewer.less
… awx/ui/client/legacy/styles/fonts.less
… awx/ui/client/legacy/styles/forms.less
… awx/ui/client/legacy/styles/inventory-edit.less
… awx/ui/client/legacy/styles/job-details.less
… awx/ui/client/legacy/styles/jobs.less
… awx/ui/client/legacy/styles/jquery-ui-overrides.less
… awx/ui/client/legacy/styles/lists.less
… awx/ui/client/legacy/styles/log-viewer.less
… awx/ui/client/legacy/styles/stdout.less
… awx/ui/client/legacy/styles/survey-maker.less
… awx/ui/client/legacy/styles/text-label.less
… awx/ui/client/lib/components/input/text.directive.js
… awx/ui/client/lib/components/input/textarea-secret.directive.js
… awx/ui/client/lib/components/input/textarea.directive.js
… awx/ui/client/lib/components/layout/layout.directive.js
… awx/ui/client/lib/components/layout/side-nav-item.directive.js
… awx/ui/client/lib/components/layout/side-nav.directive.js
… awx/ui/client/lib/components/layout/top-nav-item.directive.js
… awx/ui/client/lib/components/modal/modal.directive.js
… awx/ui/client/lib/components/panel/body.directive.js
… awx/ui/client/lib/components/panel/heading.directive.js