* Faster build times
* Smaller bundle sizes
* Adjust paths
* Cleanup npm dependencies
* Remove unneded Grunt tasks

Comparison:

Item | Old | New
------------ | ------------- | -------------
Build Time (dev) | 1m15.4s | 23.5s
vendor.js size | 13.9 MB | 4.4 MB
app.js size | 8.7 MB | 4.1 MB
vendor.css size | 200 KB | 194 KB
app.css size | 293 KB | 271 KB

Many of the file changes are due to small path updates and some minor restructuring. Full compatibility should be retained for the existing Make and npm targets.

Note: In order to leverage Webpack's dev server, `templates/ui/index.html` is generated from `client/src/index.template.html`. 

To test out this change:

* Remove `node_modules`
* `npm i` (from `awx/ui`)
* Use one of the following:
  * `npm run dev` for dev build
  * `npm run watch` for continuous dev (@ `http://localhost:3000`)
... Or if you prefer the make targets from the project root:
  * `make ui-devel`
  * `make ui-docker`