@jakemcdermott 2017-09-08
I gave this a whirl by running:

```
make clean-ui
npm --preix awx/ui install
npm --prefix awx/ui run watch
```

From there I changed various files to check out the syncing behavior and everything seems to be working smoothly. 👍