@gconsidine 2017-09-08
Cool, thanks for taking a look guys. I want to take another stab at resolving the extraneous `/` in the static file paths this morning before bringing this one in.